Add new commands from dev to RespCommandHashLookup:
- Vector Set: VADD, VCARD, VDIM, VEMB, VGETATTR, VINFO, VISMEMBER, VLINKS,
  VRANDMEMBER, VREM, VSETATTR, VSIM
- Range Index (dot-prefixed wire names): RI.CREATE, RI.SET, RI.GET, RI.DEL,
  RI.RANGE, RI.SCAN, RI.EXISTS, RI.CONFIG, RI.METRICS
- String: SETWITHETAG
- Cluster subcommands: rename AOFSYNC -> ADVANCE_TIME, add MLOG_KEY_TIME, RESERVE
- Internal-only RIPROMOTE, RIRESTORE never come from wire (no hash entry needed)

+++ libs/cluster/Server/Migration/MigrateOperation.cs
@@ -2,13 +2,18 @@
 // Licensed under the MIT license.
 
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.Threading.Tasks;
 using Garnet.client;
 using Garnet.server;
+using Microsoft.Extensions.Logging;
+using Tsavorite.core;
 
 namespace Garnet.cluster
 {
-    internal sealed unsafe partial class MigrateSession : IDisposable
+    internal sealed partial class MigrateSession : IDisposable
     {
         internal sealed partial class MigrateOperation
         {
@@ -16,15 +21,30 @@ namespace Garnet.cluster
             public readonly List<byte[]> keysToDelete;
             public StoreScan storeScan;
 
+            private readonly ConcurrentDictionary<byte[], byte[]> vectorSetsIndexKeysToMigrate;
+
             readonly MigrateSession session;
             readonly GarnetClientSession gcs;
             readonly LocalServerSession localServerSession;
 
             public GarnetClientSession Client => gcs;
 
+            public IEnumerable<KeyValuePair<byte[], byte[]>> VectorSets => vectorSetsIndexKeysToMigrate;
+
             public void ThrowIfCancelled() => session._cts.Token.ThrowIfCancellationRequested();
 
             public bool Contains(int slot) => session._sslots.Contains(slot);
+
+            public bool ContainsNamespace(ReadOnlySpan<byte> namespaceBytes)
+            {
+                Debug.Assert(namespaceBytes.Length == 1, "Longer namespaces note supported");
+
+                var ns = (ulong)namespaceBytes[0];
+
+                return session._namespaces?.Contains(ns) ?? false;
+            }
+            public void EncounteredVectorSet(byte[] key, byte[] value)
+            => vectorSetsIndexKeysToMigrate.TryAdd(key, value);
 
             public MigrateOperation(MigrateSession session, Sketch sketch = null, int batchSize = 1 << 18)
             {
@@ -34,11 +54,12 @@ namespace Garnet.cluster
                 this.sketch = sketch ?? new(keyCount: batchSize << 2);
                 storeScan = new StoreScan(this);
                 keysToDelete = [];
+                vectorSetsIndexKeysToMigrate = new(ByteArrayComparer.Instance);
             }
 
-            public bool Initialize()
+            public async ValueTask<bool> InitializeAsync()
             {
-                if (!session.CheckConnection(gcs))
+                if (!await session.CheckConnectionAsync(gcs).ConfigureAwait(false))
                     return false;
                 gcs.InitializeIterationBuffer(session.clusterProvider.storeWrapper.loggingFrequency);
                 return true;
@@ -63,39 +84,60 @@ namespace Garnet.cluster
             /// Transmit gathered keys
             /// </summary>
             /// <returns></returns>
-            public bool TransmitSlots()
+            public async Task<bool> TransmitSlotsAsync()
             {
-                var output = new UnifiedOutput();    // TODO: initialize this based on gcs curr and end; make sure it has the initial part of the "send" set
+                var output = new UnifiedOutput();       // TODO: initialize this based on gcs curr and end; make sure it has the initial part of the "send" set
+                var vectorOutput = new VectorOutput();  // TODO: initialize this based on gcs curr and end; make sure it has the initial part of the "send" set
 
                 try
                 {
                     var input = new UnifiedInput(RespCommand.MIGRATE);
                     input.arg1 = session.NetworkBufferSettings.sendBufferSize - common.NetworkBufferSettings.SendBufferOverheadReserve;
-                    foreach (var key in sketch.argSliceVector)
+
+                    VectorInput vectorInput = new();
+                    vectorInput.AlignmentExpected = true; // We're moving DiskANN sourced data, so alignment is expected
+                    vectorInput.MaxMigrationHeapAllocationSize = session.NetworkBufferSettings.sendBufferSize - common.NetworkBufferSettings.SendBufferOverheadReserve;
+
+                    foreach (var (ns, key, hasNs) in sketch.argSliceVector)
                     {
-                        if (!session.WriteOrSendRecord(gcs, localServerSession, key, ref input, ref output, out _))
-                            return false;
+                        if (hasNs)
+                        {
+                            // Migrating Vector Set element data
+                            if (!await session.WriteOrSendRecordAsync(gcs, localServerSession, ns, key, ref vectorInput, ref vectorOutput, out _).ConfigureAwait(false))
+                                return false;
+                        }
+                        else
+                        {
+                            // Migrating everything else
+                            if (!await session.WriteOrSendRecordAsync(gcs, localServerSession, key, ref input, ref output, out _).ConfigureAwait(false))
+                                return false;
+                        }
                     }
 
                     // Flush final data in client buffer
-                    if (!session.HandleMigrateTaskResponse(gcs.SendAndResetIterationBuffer()))
+                    if (!await session.HandleMigrateTaskResponseAsync(gcs.SendAndResetIterationBuffer()).ConfigureAwait(false))
                         return false;
                 }
                 finally
                 {
                     output.SpanByteAndMemory.Dispose();
+                    vectorOutput.SpanByteAndMemory.Dispose();
                 }
 
                 return true;
             }
 
-            public bool TransmitKeys()
+            public async Task<bool> TransmitKeysAsync(Dictionary<byte[], byte[]> vectorSetKeysToIgnore)
             {
                 // Use this for both stores; main store will just use the SpanByteAndMemory directly. We want it to be outside iterations
                 // so we can reuse the SpanByteAndMemory.Memory across iterations.
                 // TODO: initialize 'output' based on gcs curr and end; make sure it has the initial part of the "send" set, and call gcs.IncrementRecordDirect().
                 //       This will still allow SBAM.Memory to be reused.
                 var output = new UnifiedOutput();
+
+#if NET9_0_OR_GREATER
+                var ignoreLookup = vectorSetKeysToIgnore.GetAlternateLookup<ReadOnlySpan<byte>>();
+#endif
 
                 try
                 {
@@ -110,7 +152,23 @@ namespace Garnet.cluster
                         if (keys[i].Item2)
                             continue;
 
-                        if (!session.WriteOrSendRecord(gcs, localServerSession, keys[i].Item1, ref input, ref output, out var status))
+                        var spanByte = keys[i].Item1;
+
+                        // Don't transmit if a Vector Set
+                        var isVectorSet =
+                            vectorSetKeysToIgnore.Count > 0 &&
+#if NET9_0_OR_GREATER
+                            ignoreLookup.ContainsKey(spanByte.ReadOnlySpan);
+#else
+                                vectorSetKeysToIgnore.ContainsKey(spanByte.ToArray());
+#endif
+
+                        if (isVectorSet)
+                        {
+                            continue;
+                        }
+
+                        if (!await session.WriteOrSendRecordAsync(gcs, localServerSession, keys[i].Item1, ref input, ref output, out var status).ConfigureAwait(false))
                             return false;
 
                         // If key was FOUND, mark it for deletion
@@ -119,13 +177,65 @@ namespace Garnet.cluster
                     }
 
                     // Flush final data in client buffer
-                    if (!session.HandleMigrateTaskResponse(gcs.SendAndResetIterationBuffer()))
+                    if (!await session.HandleMigrateTaskResponseAsync(gcs.SendAndResetIterationBuffer()).ConfigureAwait(false))
                         return false;
                 }
                 finally
                 {
                     output.SpanByteAndMemory.Dispose();
                 }
+                return true;
+            }
+
+            /// <summary>
+            /// Transmit data in namespaces during a MIGRATE ... KEYS operation.
+            /// 
+            /// Doesn't delete anything, just scans and transmits.
+            /// </summary>
+            public async ValueTask<bool> TransmitKeysNamespacesAsync(ILogger logger)
+            {
+                var migrateOperation = this;
+
+                if (!await migrateOperation.InitializeAsync().ConfigureAwait(false))
+                    return false;
+
+                var workerStartAddress = migrateOperation.session.clusterProvider.storeWrapper.store.Log.BeginAddress;
+                var workerEndAddress = migrateOperation.session.clusterProvider.storeWrapper.store.Log.TailAddress;
+
+                var cursor = workerStartAddress;
+                logger?.LogWarning("<MainStore> migrate keys (namespaces) scan range [{workerStartAddress}, {workerEndAddress}]", workerStartAddress, workerEndAddress);
+                while (true)
+                {
+                    var current = cursor;
+                    // Build Sketch
+                    migrateOperation.sketch.SetStatus(SketchStatus.INITIALIZING);
+                    migrateOperation.Scan(ref current, workerEndAddress);
+
+                    // Stop if no keys have been found
+                    if (migrateOperation.sketch.argSliceVector.IsEmpty) break;
+
+                    logger?.LogWarning("Scan from {cursor} to {current} and discovered {count} keys", cursor, current, migrateOperation.sketch.argSliceVector.Count);
+
+                    // Transition EPSM to MIGRATING
+                    migrateOperation.sketch.SetStatus(SketchStatus.TRANSMITTING);
+                    await migrateOperation.session.WaitForConfigPropagationAsync().ConfigureAwait(false);
+
+                    // Transmit all keys gathered
+                    if (!await migrateOperation.TransmitSlotsAsync().ConfigureAwait(false))
+                    {
+                        logger?.LogWarning("TransmitSlots failed for {cursor} to {current} (with {count} keys)", cursor, current, migrateOperation.sketch.argSliceVector.Count);
+                        return false;
+                    }
+
+                    // Transition EPSM to DELETING
+                    migrateOperation.sketch.SetStatus(SketchStatus.DELETING);
+                    await migrateOperation.session.WaitForConfigPropagationAsync().ConfigureAwait(false);
+
+                    // Clear keys from buffer
+                    migrateOperation.sketch.Clear();
+                    cursor = current;
+                }
+
                 return true;
             }
 
@@ -138,8 +248,18 @@ namespace Garnet.cluster
                     return;
                 if (session.transferOption == TransferOption.SLOTS)
                 {
-                    foreach (var key in sketch.argSliceVector)
-                        _ = localServerSession.BasicGarnetApi.DELETE(key);
+                    foreach (var (ns, key, hasNs) in sketch.argSliceVector)
+                    {
+                        if (hasNs)
+                        {
+                            // Namespace'd keys are deleted as part after migration completes
+                            continue;
+                        }
+                        else
+                        {
+                            _ = localServerSession.BasicGarnetApi.DELETE(key);
+                        }
+                    }
                 }
                 else
                 {
@@ -151,6 +271,19 @@ namespace Garnet.cluster
                             _ = localServerSession.BasicGarnetApi.DELETE(keys[i].Item1);
                     }
                 }
+            }
+
+            /// <summary>
+            /// Delete a Vector Set after migration if _copyOption is not set.
+            /// </summary>
+            public void DeleteVectorSet(PinnedSpanByte key)
+            {
+                if (session._copyOption)
+                    return;
+
+                var delRes = localServerSession.BasicGarnetApi.DELETE(key);
+
+                session.logger?.LogDebug("Deleting Vector Set {key} after migration: {delRes}", System.Text.Encoding.UTF8.GetString(key), delRes);
             }
         }
     }
