+++ libs/cluster/Server/Migration/MigrateSessionSlots.cs
@@ -2,21 +2,78 @@
 // Licensed under the MIT license.
 
 using System;
+using System.Buffers;
+using System.Buffers.Binary;
+using System.Collections.Frozen;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
 using System.Threading.Tasks;
+using Garnet.client;
+using Garnet.server;
+
 #if DEBUG
 using Garnet.common;
 #endif
+
 using Microsoft.Extensions.Logging;
+using Tsavorite.core;
 
 namespace Garnet.cluster
 {
     internal sealed partial class MigrateSession : IDisposable
     {
         /// <summary>
+        /// Attempts to reserve contexts on the destination node for migrating vector sets.
+        /// 
+        /// This maps roughly to "for each <see cref="VectorManager.ContextStep"/> namespaces, reserve one context, record the mapping".
+        /// </summary>
+        public async Task<bool> ReserveDestinationVectorSetsAsync()
+        {
+            Debug.Assert((_namespaces.Count % (int)VectorManager.ContextStep) == 0, "Expected to be migrating Vector Sets, and thus to have an even number of namespaces");
+
+            var neededContexts = _namespaces.Count / (int)VectorManager.ContextStep;
+
+            try
+            {
+                var reservedCtxs = await migrateOperation[0].Client.ExecuteForArrayAsync("CLUSTER", "RESERVE", "VECTOR_SET_CONTEXTS", neededContexts.ToString());
+
+                var rootNamespacesMigrating = _namespaces.Where(static x => (x % VectorManager.ContextStep) == 0);
+
+                var nextReservedIx = 0;
+
+                var namespaceMap = new Dictionary<ulong, ulong>();
+
+                foreach (var migratingContext in rootNamespacesMigrating)
+                {
+                    var toMapTo = ulong.Parse(reservedCtxs[nextReservedIx]);
+                    for (var i = 0U; i < VectorManager.ContextStep; i++)
+                    {
+                        var fromCtx = migratingContext + i;
+                        var toCtx = toMapTo + i;
+
+                        namespaceMap[fromCtx] = toCtx;
+                    }
+
+                    nextReservedIx++;
+                }
+
+                _namespaceMap = namespaceMap.ToFrozenDictionary();
+
+                return true;
+            }
+            catch (Exception ex)
+            {
+                logger?.LogError(ex, "Failed to reserve {count} Vector Set contexts on destination node {node}", neededContexts, _targetNodeId);
+                return false;
+            }
+        }
+
+        /// <summary>
         /// Migrate Slots inline driver
         /// </summary>
         /// <returns></returns>
-        public async Task<bool> MigrateSlotsDriverInline()
+        public async Task<bool> MigrateSlotsDriverInlineAsync()
         {
             var storeBeginAddress = clusterProvider.storeWrapper.store.Log.BeginAddress;
             var storeTailAddress = clusterProvider.storeWrapper.store.Log.TailAddress;
@@ -29,38 +86,122 @@ namespace Garnet.cluster
 
             // Send store
             logger?.LogWarning("Store migrate scan range [{storeBeginAddress}, {storeTailAddress}]", storeBeginAddress, storeTailAddress);
-            var success = await CreateAndRunMigrateTasks(storeBeginAddress, storeTailAddress, storePageSize);
+
+            var success = await CreateAndRunMigrateTasksAsync(storeBeginAddress, storeTailAddress, storePageSize);
             if (!success) return false;
 
             return true;
 
-            async Task<bool> CreateAndRunMigrateTasks(long beginAddress, long tailAddress, int pageSize)
+            async Task<bool> CreateAndRunMigrateTasksAsync(long beginAddress, long tailAddress, int pageSize)
             {
-                logger?.LogTrace("{method} > Scan in range ({BeginAddress},{TailAddress})", nameof(CreateAndRunMigrateTasks), beginAddress, tailAddress);
-                var migrateOperationRunners = new Task[clusterProvider.serverOptions.ParallelMigrateTaskCount];
+                logger?.LogTrace("{method} > Scan in range ({BeginAddress},{TailAddress})", nameof(CreateAndRunMigrateTasksAsync), beginAddress, tailAddress);
+                var migrateOperationRunners = new Task<bool>[clusterProvider.serverOptions.ParallelMigrateTaskCount];
+
                 var i = 0;
                 while (i < migrateOperationRunners.Length)
                 {
                     var idx = i;
-                    migrateOperationRunners[idx] = Task.Run(() => ScanStoreTask(idx, beginAddress, tailAddress, pageSize));
+                    migrateOperationRunners[idx] = ScanStoreTaskAsync(idx, beginAddress, tailAddress, pageSize);
                     i++;
                 }
 
                 try
                 {
-                    await Task.WhenAll(migrateOperationRunners).WaitAsync(_timeout, _cts.Token).ConfigureAwait(false);
+                    var scanResults = await Task.WhenAll(migrateOperationRunners).WaitAsync(_timeout, _cts.Token).ConfigureAwait(false);
+                    if (!scanResults.All(static x => x))
+                    {
+                        logger?.LogWarning("Aborting migration due to ScanStoreTask failure");
+                        return false;
+                    }
+
+                    // Handle migration of discovered Vector Set keys now that they're namespaces have been moved
+                    var vectorSets = migrateOperation.SelectMany(static mo => mo.VectorSets).GroupBy(static g => g.Key, ByteArrayComparer.Instance).ToDictionary(static g => g.Key, g => g.First().Value, ByteArrayComparer.Instance);
+
+                    if (vectorSets.Count > 0)
+                    {
+                        var gcs = migrateOperation[0].Client;
+
+                        var serializeBufferArr = ArrayPool<byte>.Shared.Rent(128);
+                        try
+                        {
+                            foreach (var (key, value) in vectorSets)
+                            {
+                                // Update the index context as we move it, so it arrives on the destination node pointed at the appropriate
+                                // namespaces for element data
+                                VectorManager.ReadIndex(value, out var oldContext, out _, out _, out _, out _, out _, out _, out _, out _);
+
+                                var newContext = _namespaceMap[oldContext];
+                                VectorManager.SetContextForMigration(value, newContext);
+
+                                var neededSpace = sizeof(int) + key.Length + sizeof(int) + value.Length;
+
+                                if (neededSpace > serializeBufferArr.Length)
+                                {
+                                    ArrayPool<byte>.Shared.Return(serializeBufferArr);
+                                    serializeBufferArr = ArrayPool<byte>.Shared.Rent(neededSpace);
+                                }
+
+                                // Scope so Span doesn't cross await boundary
+                                {
+                                    Span<byte> serializeBuffer = serializeBufferArr;
+                                    BinaryPrimitives.WriteInt32LittleEndian(serializeBuffer, key.Length);
+                                    key.CopyTo(serializeBuffer[sizeof(int)..]);
+                                    BinaryPrimitives.WriteInt32LittleEndian(serializeBuffer[(sizeof(int) + key.Length)..], value.Length);
+                                    value.CopyTo(serializeBuffer[(sizeof(int) + key.Length + sizeof(int))..]);
+                                }
+
+                                if (gcs.NeedsInitialization)
+                                    gcs.SetClusterMigrateHeader(_sourceNodeId, _replaceOption, isVectorSets: true);
+
+                                while (!gcs.TryWriteRecordSpan(serializeBufferArr.AsSpan()[..neededSpace], MigrationRecordSpanType.VectorSetIndex, out var task))
+                                {
+                                    if (!await HandleMigrateTaskResponseAsync(task).ConfigureAwait(false))
+                                    {
+                                        logger?.LogCritical("Failed to migrate Vector Set key {key} during migration", SpanByte.ToShortString(key));
+                                        return false;
+                                    }
+
+                                    gcs.SetClusterMigrateHeader(_sourceNodeId, _replaceOption, isVectorSets: true);
+                                }
+
+                                // Force a flush before doing the delete, in case that fails
+                                if (!await HandleMigrateTaskResponseAsync(gcs.SendAndResetIterationBuffer()).ConfigureAwait(false))
+                                {
+                                    logger?.LogCritical("Flush failed before deletion of Vector Set {key} duration migration", SpanByte.ToShortString(key));
+                                    return false;
+                                }
+
+                                // Delete the index on this node now that it's moved over to the destination node
+                                unsafe
+                                {
+                                    fixed (byte* keyPtr = key)
+                                    {
+                                        var pinnedKeySpan = PinnedSpanByte.FromPinnedPointer(keyPtr, key.Length);
+                                        migrateOperation[0].DeleteVectorSet(pinnedKeySpan);
+                                    }
+                                }
+                            }
+                        }
+                        finally
+                        {
+                            ArrayPool<byte>.Shared.Return(serializeBufferArr);
+                        }
+                    }
                 }
                 catch (Exception ex)
                 {
-                    logger?.LogError(ex, "{CreateAndRunMigrateTasks}: {beginAddress} {tailAddress} {pageSize}", nameof(CreateAndRunMigrateTasks), beginAddress, tailAddress, pageSize);
-                    _cts.Cancel();
+                    logger?.LogError(ex, "{CreateAndRunMigrateTasks}: {beginAddress} {tailAddress} {pageSize}", nameof(CreateAndRunMigrateTasksAsync), beginAddress, tailAddress, pageSize);
+                    await _cts.CancelAsync().ConfigureAwait(false);
                     return false;
                 }
                 return true;
             }
 
-            Task<bool> ScanStoreTask(int taskId, long beginAddress, long tailAddress, int pageSize)
+            async Task<bool> ScanStoreTaskAsync(int taskId, long beginAddress, long tailAddress, int pageSize)
             {
+                // Force async
+                await Task.Yield();
+
                 var migrateOperation = this.migrateOperation[taskId];
                 var range = (tailAddress - beginAddress) / clusterProvider.storeWrapper.serverOptions.ParallelMigrateTaskCount;
                 var workerStartAddress = beginAddress + (taskId * range);
@@ -68,8 +209,8 @@ namespace Garnet.cluster
 
                 workerStartAddress = workerStartAddress - (2 * pageSize) > 0 ? workerStartAddress - (2 * pageSize) : 0;
                 workerEndAddress = workerEndAddress + (2 * pageSize) < storeTailAddress ? workerEndAddress + (2 * pageSize) : storeTailAddress;
-                if (!migrateOperation.Initialize())
-                    return Task.FromResult(false);
+                if (!await migrateOperation.InitializeAsync().ConfigureAwait(false))
+                    return false;
 
                 var cursor = workerStartAddress;
                 logger?.LogWarning("<{taskId}> migrate scan range [{workerStartAddress}, {workerEndAddress}]", taskId, workerStartAddress, workerEndAddress);
@@ -88,14 +229,18 @@ namespace Garnet.cluster
 
                     // Transition EPSM to MIGRATING
                     migrateOperation.sketch.SetStatus(SketchStatus.TRANSMITTING);
-                    WaitForConfigPropagation();
+                    await WaitForConfigPropagationAsync().ConfigureAwait(false);
 
                     // Transmit all keys gathered
-                    migrateOperation.TransmitSlots();
+                    if (!await migrateOperation.TransmitSlotsAsync().ConfigureAwait(false))
+                    {
+                        logger?.LogWarning("[{taskId}> TransmitSlots failed for {cursor} to {current} (with {count} keys)", taskId, cursor, current, migrateOperation.sketch.argSliceVector.Count);
+                        return false;
+                    }
 
                     // Transition EPSM to DELETING
                     migrateOperation.sketch.SetStatus(SketchStatus.DELETING);
-                    WaitForConfigPropagation();
+                    await WaitForConfigPropagationAsync().ConfigureAwait(false);
 
                     // Deleting keys (Currently gathering keys from push-scan and deleting them outside)
                     migrateOperation.DeleteKeys();
@@ -105,7 +250,7 @@ namespace Garnet.cluster
                     cursor = current;
                 }
 
-                return Task.FromResult(true);
+                return true;
             }
         }
     }
